bon le checkerboard etais fait pour le plane mais pour le rendre mieux g tenter des resolutions d equations matriciel, mais g chierQQpart donc triste

+++ submodules/backend-raytracer/include/rt/render/backend/raytracer/ray.h
@@ -6,7 +6,7 @@
 /*   By: kiroussa <oss@xtrm.me>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/14 08:51:32 by kiroussa          #+#    #+#             */
-/*   Updated: 2024/10/25 00:55:38 by yroussea         ###   ########.fr       */
+/*   Updated: 2024/10/25 23:41:13 by yroussea         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,9 +29,9 @@ typedef struct s_ray
 	t_color	color;
 }	t_ray;
 
-void	eye_rays(t_ray *ray, float u, float v, float fov);
-float	get_width(t_rt_backend *backend, float x);
-float	get_height(t_rt_backend *backend, float y);
+void	eye_rays(t_ray *ray, double u, double v, double fov);
+double	get_width(t_rt_backend *backend, double x);
+double	get_height(t_rt_backend *backend, double y);
 
 # endif // __RT_RENDER_BACKEND_RAYTRACER_RAY_H__
 #endif // RAY_H
